Address GPT re-review findings
- ConfigValue::of throws logic errors for out-of-range unsigned and
  null C strings in all builds; ConfigModel::set checks both first so
  the Result API never throws (factories map to MigrationFailed)
- MigrationEngine invokes the stored callable directly, restoring
  stateful-callback semantics; mutating the registry mid-migration is
  documented as forbidden
- Scalar range check compares in long double: get<long double>() no
  longer narrows the target type's limits (UB)
- XML save detects pugixml set_name allocation failure on named
  appends
- Fetched tl-expected stays out of the exported link interface (its
  headers ship in our includedir), fixing custom-includedir installs;
  find_dependency(pugixml) runs only when the xml component is wanted
- Document top-level-dependent test/example defaults and installed-
  package dependency behavior

+++ include/configmanager/config_node.hpp
@@ -103,10 +103,14 @@ Result<T> convertScalar(const Scalar& scalar) {
         return static_cast<T>(*value);  // NaN converts to NaN losslessly
       }
       // Out-of-range finite values must not reach the cast: narrowing a
-      // double outside T's finite range is undefined behavior.
+      // double outside T's finite range is undefined behavior. Compared in
+      // long double so T's limits only ever widen (converting a wider T's
+      // limits to double would itself be out of range).
       if (std::isfinite(*value) &&
-          (*value < static_cast<double>(std::numeric_limits<T>::lowest()) ||
-           *value > static_cast<double>(std::numeric_limits<T>::max()))) {
+          (static_cast<long double>(*value) <
+               static_cast<long double>(std::numeric_limits<T>::lowest()) ||
+           static_cast<long double>(*value) >
+               static_cast<long double>(std::numeric_limits<T>::max()))) {
         return fail(ErrorCode::InvalidType,
                     "Double value is outside the requested type's range");
       }
